requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

dnspython==2.6.1
